requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
python-dotenv
pydantic
requests
duckduckgo-search 
langchain-community
faiss-cpu 
pypdf
pandas 
pillow
//sentence-transformers
langchain-experimental
fastembed